Unify error handling from api responses Fixes #130
